--- a/gdxpds/gdx.py
+++ b/gdxpds/gdx.py
@@ -805,15 +805,16 @@ class GdxSymbol(object):
             self._loaded = True
             return
 
-        data = []
         _ret, records = gdxcc.gdxDataReadStrStart(self.file.H,self.index)
-        for _i in range(records):
-            _ret, elements, values, _afdim = gdxcc.gdxDataReadStr(self.file.H)
-            # make sure we pick value columns up correctly
-            data.append(elements + [values[col_ind] for col_name, col_ind in self.value_cols])
-            if self.data_type == GamsDataType.Set:
-                data[-1][-1] = True
-                # gdxdict called gdxGetElemText here, but I do not currently see value in doing that
+
+        def reader():
+            handle = self.file.H
+            for i in range(records):
+                yield gdxcc.gdxDataReadStr(handle)
+
+        vc = self.value_cols  # do this for speed in the next line
+        data = [elements + [values[col_ind] for col_name, col_ind in vc] for ret, elements, values, afdim in reader()]
+        # gdxdict called gdxGetElemText here, but I do not currently see value in doing that
         self.dataframe = data
         if not self.data_type == GamsDataType.Set:
             self.dataframe = special.convert_gdx_to_np_svs(self.dataframe, self.num_dims)